+++ apps/ui/src/lib/components/SpeciesDetailModal.svelte
@@ -594,6 +594,7 @@
                                                     }))}
                                                 userLocation={authStore.canModify && settingsStore.settings?.location_latitude && settingsStore.settings?.location_longitude ? [settingsStore.settings.location_latitude, settingsStore.settings.location_longitude] : null}
                                                 zoom={10}
+                                                obfuscate={!authStore.canModify}
                                             />
                                         </div>
                                     {/if}